dibbler: update to latest upstream version + misc recipe cleanup
Set PV based on SRCPV (ie abbreviated git rev) instead of SRCREV.

Disable resolvconf PACKAGECONFIG by default. Since there's no runtime
dependency on resolvconf, enabling support in dibbler isn't going to
achieve much on its own. Note that if / when a runtime dependency on
resolvconf is added, it should be added to the specific dibbler
package(s) which need it (ie only the client?), not to all of them.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>