dibbler: force static linking with libstdc++
Dibbler is written in C++, therefore by default the dibbler packages
each have a runtime dependency on libstdc++, which means an ~1MB
library on the target.

If only one of the dibbler packages is required and no other C++
based applications are going to be included in the image then linking
the dibbler apps statically with libstdc++ will save some space in
the final rootfs image.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>