lighttpd: drop dependency on openssl MD4 support
Fix for build errors in mod_authn_file.c when MD4 support is disabled
in openssl.

The MD4 dependency was added in approx Aug 2016 as part of a change
to add support for the CRYPT-MD5-NTLM password hashing algorithm:

  https://redmine.lighttpd.net/projects/lighttpd/repository/revisions/65efc2eda8419e4bba016df1b77783bb55c2b269/diff/src/mod_authn_file.c
  https://redmine.lighttpd.net/issues/1743
  https://redmine.lighttpd.net/projects/lighttpd/wiki/Docs_ModAuth

It seems unlikely that this will be required for an embedded system,
so looks safe to disable.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>