iptables: force static libs to be compiled with -fPIC
Ensure that the iptables static libs can be safely linked into a PIE
application as a workaround for issues building iproute2 with clang
(which defaults to PIE).

This workaround can be dropped if either iptables is configured to
provide shared libs or iproute2 is configured to avoid linking with
iptables.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>