glibc: prefer glibc RPC if available (OE 2.5 and earlier)
The glibc recipes in OE 2.5 and earlier provide support for RPC by
default:

  OE 2.2 : glibc 2.24 : RPC supported
  OE 2.3 : glibc 2.25 : RPC supported
  OE 2.4 : glibc 2.26 : RPC optional (enabled by OE)
  OE 2.5 : glibc 2.27 : RPC optional (enabled by OE)
  OE 2.6 : glibc 2.28 : RPC optional (disabled by OE)
  OE 2.7 : glibc 2.29 : RPC optional (disabled by OE)
  OE 3.0 : glibc 2.30 : RPC optional (disabled by OE)
  OE 3.1 : glibc 2.31 : RPC optional (disabled by OE)

For OE 2.2 and later (ie glibc 2.24, 2.25, 2.26 and 2.27), set glibc
as the prefered provider for libucrpc.

Builds based on musl or versions of glibc for which RPC support is
disabled by default in oe-core will continue to use the standalone
libucrpc.

  https://savannah.gnu.org/forum/forum.php?forum_id=8921

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>