msgpack-c: import recipe for msgpack-c 2.1.5 from meta-oe
Unmodified snapshot from meta-oe OE 2.5 (sumo):

  https://git.openembedded.org/meta-openembedded/commit/?id=cebc29d941b12a958adf5f41822534e344472650

Pick the last version of msgpack-c from the 2.1.x series since 3.0.0
has an API change. It looks minor but don't risk updating for now
(RDK is developed and tested with msgpack-c 2.1.1).

  https://github.com/msgpack/msgpack-c/releases

Use the 2.1.5 recipe from OE 2.5 (not the one from OE 2.6) since it's
the last one which explicitly packages ${libdir}/cmake (newer
versions of the recipe rely on cmake.bbclass to do that, but we may
want to use older versions of oe-core which don't have that update to
cmake.bbclass).

  http://git.openembedded.org/openembedded-core/commit/?id=d91dc4666683a96e9d03cbbd21b8a546f9069c93

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>