setup-environment: hide machine configs from meta-cmf-raspberrypi, meta-intel and meta-intelce
There are a lot of machine configs in meta-cmf-raspberrypi. Since we
don't ever expect to build these machines with the mng distro, hide
them when picking a machine interactively to avoid confusion.

Hide machine configs in meta-intel (ie the generic Intel open source
meta layer) and meta-intelce (ie an Intel proprietary meta layer in
the Puma 6 / Puma 7 SDKs) for a similar reason. Note that a single
pattern is used to match both meta-intel and meta-intelce.

As always, it's still possible to select any valid machine manually.
This change is only related to the list of machines presented to the
user when running the setup-environment script interactively.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>