mng.conf: ensure KERNEL_PACKAGE_NAME is defined
The KERNEL_PACKAGE_NAME variable was added in OE 2.5 (sumo). If it's
not defined then RDEPENDS_${KERNEL_PACKAGE_NAME}-base doesn't map to
RDEPENDS_kernel-base, which is the variable used in OE 2.4 and
earlier to control whether the kernel image gets included in the
rootfs.

For OE 2.5 and later KERNEL_PACKAGE_NAME has a default value set from
kernel.bbclass but setting it explicitly should be fine too.

  https://git.openembedded.org/openembedded-core/commit/?id=6c8c899849d101fd1b86aad0b8eed05c7c785924

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>