base-files: drop support for persistent /var/log for compatibility with OE 2.3 and earlier
Support for VOLATILE_LOG_DIR was first added in OE 2.4 and relies on
changes in various recipes (not only base-files). If VOLATILE_LOG_DIR
is set to "yes" (the default) then /var/log will be volatile (ie in
tmpfs and not persistent across reboots). If VOLATILE_LOG_DIR is set
to "no" then /var/log will be normal directory within the rootfs.

Before the VOLATILE_LOG_DIR variable was added, /var/log was always
volatile (ie the same as the default behaviour in OE 2.4 and later).

Since for Puma 6 ARM we never expect to have persistent /var.log (and
there's nowhere in Flash for it anyway), we can remove support for
the VOLATILE_LOG_DIR variable to make this base-files recipe
compatible with versions of OE prior to 2.4.

This commit effectively reverts the following change from oe-core:

  https://git.openembedded.org/openembedded-core/commit/?id=fc19114ca409fc3178ae5cbb44db875b12eeedfc

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>