e2fsprogs: explicitly disable systemd support
If systemd is auto detected, then various service files will be
installed, which are not covered by the existing packaging rules:

 | WARNING: e2fsprogs-1.45.4-r0 do_package: QA Issue: e2fsprogs: Files/directories were installed but not shipped in any package:
 |   /lib/systemd
 |   /lib/e2fsprogs
 |   /lib/systemd/system
 |   /lib/systemd/system/e2scrub_reap.service
 |   /lib/systemd/system/e2scrub_all.service
 |   /lib/systemd/system/e2scrub@.service
 |   /lib/systemd/system/e2scrub_all.timer
 |   /lib/systemd/system/e2scrub_fail@.service
 |   /lib/e2fsprogs/e2scrub_fail
 | Please set FILES such that these items are packaged. Alternatively if they are unneeded, avoid installing them or delete them within do_install.
 | e2fsprogs: 9 installed and not shipped files. [installed-vs-shipped]

This issue is presumably only seen when a newer version of e2fsprogs
is backported to an older version of OE without support for Recipe
Specific Sysroots (ie OE 2.2 and earlier). The above warning comes
from OE 2.2.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>