base-files: cleanup and simplify /etc/profile
Drop generic comments and make the assumption that /etc/profile is
responsible for setting variables such as PS1, EDITOR and TERM (ie
have /etc/profile always set these variables rather than the original
behaviour of just providing backup values in the case if they are not
set elsewhere).

Note that PS1 is only set if a value is already present.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>