mng.conf: clear LDCONFIGDEPEND ( prevent creation of /etc/ld.so.cache )
Clear LDCONFIGDEPEND to avoid openembedded-core/meta/lib/oe/rootfs.py
running ldconfig to create /etc/ld.so.cache as part of creating the
rootfs.

This should really be fixed upstream so that running ldconfig is
controlled by the ldconfig distro feature, but until then we need to
clear LDCONFIGDEPEND explicitly.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>