curl: switch from gnutls -> openssl
For various reasons (licensing, performance, likelihood of being a
shared dependency, etc) openssl is a better choice for most embedded
systems than gnutls. It's still a bit of a mystery why the default
configuration for the curl recipe in oe-core is gnutls...

Also disable verbose errors and dependency on libidn2.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>