base-files: mount tmpfs on /var ( not only on /var/volatile )
The Puma6 ARM SDK makes various assumptions that /var is writeable.
It's a reasonable thing to do - similar to the systemd.volatile=state
option, first added to systemd in v233:

  https://github.com/systemd/systemd/blob/v233/NEWS#L285

However, by default oe-core assumes that for read-only rootfs, /var
is read-only and only /var/volatile is writeable. Therefore, up to
now, making /var writeable has been achieved by a hack in the sdkinit
script to unmount /var/volatile and manually mount tmpfs on /var.

That hack can be removed if we update fstab to mount tmpfs on /var by
default (putting the change in fstab works for systemd too).

Note that anything under /var in the original rootfs will need to be
re-created again at run time (e.g. see the solution used in the Puma
6 ARM SDK to extract a "var.tar" tar file from sdkinit).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>