busybox: disable ip
The Puma 6 ARM SDK depends on the iproute2 version of the ip command
(and even patches it to support Intel specific extensions). RDKB
depends on it too. Neither of them work well with the minimal version
of the ip command provided by busybox.

Note that the busybox udhcpc config script (/etc/udhcpc.d/50default)
uses the ip command, so disabling it in busybox may cause issues for
minimal images which install busybox without also installing
iproute2.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>