busybox: move creation of /dev/shm into rcS (and out of mdev init script)
Creating /dev/shm doesn't really have anything to do with mdev.

In older versions of the mdev init script, an empty tmpfs was mounted
on /dev (before running "mdev -s" to populate it with device nodes)
and so /dev/shm needed to be created after that new tmpfs /dev
directory had been mounted.

However, since /dev is now devtmpfs (and mounted automatically by the
kernel or from a "mount -a" command in inittab) the mdev init script
no longer mounts anything over /dev, and so doesn't need to recreate
/dev/shm there either.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>