busybox-inittab: create basic set of dirs under /var
Since the entire /var directory is now tmpfs (instead of the oe-core
default of having /var be part of the rootfs and /var/volatile being
tmpfs) any directories etc under /var need to be created at run time.

Create a basic directory structure from inittab, directly after the
call to "mount -a" (which is when tmpfs is first mounted on /var).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>