licenses/bzip2: add local copy of bzip2 license
Upstream OE 3.1 oe-core has renamed the bzip2 license to bzip2-1.0.6.

That means, in order to be usable with versions of oe-core from
before and after the renaming, meta layers containing recipes which
use the bzip2 license (e.g. the busybox recipe in this meta layer)
need to provide their own copy of the license file.

Note that there are two possible approaches: provide a local copy of
the "bzip2" license file (and keep "bzip2" in LICENSE) or provide a
local copy of the "bzip2-1.0.6" license file (and update LICENSE in
the busybox recipe to match). Take the first approach since the bzip2
code in busybox is based on bzip2 1.0.4 and so a specific reference
to the "bzip2-1.0.6" license isn't really correct.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>