e2fsprogs: fix --enable-lto when cross compiling
The --enable-lto configure checks test for gcc-ar and gcc-ranlib (ie
binaries from the host) and fail if the host toolchain is old enough
that they are not found.

When cross compiling it should check for the cross compiler versions
instead. Setting to $AR and $RANLIB should archieve that, assuming
these variables are exported by the build environment (they are for
OE, but may not be in other cases).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>