dibbler: add dibbler_crash_fix.patch
The upstream meta-oe recipe already includes a patch for getSize()
but RDK has an additional fix.

Note that the patch had to be refreshed to apply cleanly on top of
the dibbler_fix_getSize_crash.patch from meta-oe (even though that
patch is present in both meta-oe and RDK, RDK seems to have a slighly
different version of it?).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>