busybox: enable cmp
The utopia service_dhcp_server.sh script currently uses diff to
determine if files are the same. It's not a very good approach. Using
cmp would be more efficient, so enable cmp in busybox to prepare for
switching the utopia scripts from diff to cmp (after which diff
should be disabled again).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>