ccsp-common-library: add explicit dependency on zlib
There are various references to zlib in ccsp-common-library. They are
mostly disabled via _NO_ANSC_ZLIB_ (which is defined via the global
ccsp CFLAGS) however the zlib.h header file is still included
unconditionally, so try to make sure it's available. The longer term
solution is to completely remove zlib support in ccsp-common-library.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>