utopia: move 15_misc symlink from post.d to registration.d
Fix for upnp service not being started in IPv4 only mode.

ccsp-provapp uses the "start-misc" event to start "service_misc",
which in turn starts the upnp service. However, in IPv4 only mode
the network is available sooner than in dual-stack mode and as a
result "start-misc" event is set sooner, before "service_misc" is
ready.

Fix by moving registration of "service_misc" to register.d, which
ensures that "service_misc" is ready before the "start-misc" event
is set.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>