dnsmasq: include fingerprint information in dhcp server leases file
Add a fingerprint field (parameter request list option) to the
dnsmasq lease file. This information is needed for Plume to be
able to identify connected stations.

Warning: this changes the format of the leases file. There is
currently no attempt to be backwards compatible with the original
file format (and unpatched versions of dnsmasq are not forward
compatible with the new file format).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>