openssl_1.0.2: enable support for ECDH
According to:

  https://www.ssllabs.com/ssltest/analyze.html?d=mv1compal.de.dmdsdp.com

The ACS requires one of the following cipher suites:

  TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384 (0xc030) ECDH secp256r1 (eq. 3072 bits RSA) FS
  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 (0xc02f) ECDH secp256r1 (eq. 3072 bits RSA) FS
  TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384 (0xc028) ECDH secp256r1 (eq. 3072 bits RSA) FS
  TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256 (0xc027) ECDH secp256r1 (eq. 3072 bits RSA) FS

To enable ECDH we need to remove "no-ecdh" from the openssl config
options (along with "no-ec", which seems to be a higher option to
disable _all_ EC based algorithms?).

Unfortunately, using the "no-ecdsa" config option on it's own seems
to cause a build failure. It needs more investigation but for now
just remove that too (ie enable ECDSA support, even though it isn't
really needed).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>