mng-image-minimal: add openssl-conf to BAD_RECOMMENDATIONS
Exclude the openssl-conf package from mng-image-minimal (and images
derived from it).

The config file isn't big but it could cause conflicts if multiple
versions of openssl are used within an image (since the path to the
config file is not version specific). The settings in openssl.cnf
seem to be mainly related to creating CA certificates etc, which
isn't relevant for use in RDKB, so try removing the conf file
entirely.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>