utopia: update udhcpc.script to support resolvconf
In the original version, $over_ride_RESOLV_CONF (ie the flag used to
determine whether resolv.conf needs to be updated and the DHCP server
restarted) was only tested in the case that /sbin/resolvconf does not
exist.

The logic needs to be inverted so that $over_ride_RESOLV_CONF is
tested first, before the check for /sbin/resolvconf.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>