utopia: update udhcpc.script to call "resolvconf -d" before "resolvconf -a"
Normally -a means "add or over-write" so running with -d first is not
required. However we have a custom patch to resolvconf to make -a
append rather than replace:

  0002-update-a-to-append-to-IFACE-rather-than-over-write.patch

The patch was added as a workaround for issues seen with dibbler
client, which calls resolvconf -a multiple times when adding
multiple DNS servers etc (whereas it should instead pass all DNS
servers into one single call to resolvconf -a ?).

The best long term approach isn't clear (fixing dibbler may be better
than patching resolvconf?) but for now update udhcpc.script to make
things work.

Fixme: to be reviewed.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>