ffmpeg: add support for auto detecting armv7ve CPUs
If --cpu= is not specified on the configure command line (or set to
"generic" as it is with recent versions of OE when using non-CPU
specific CPU tuning files, such as when building for qemuarm) the
ffmpeg configure script will try to auto detect the cpu based on the
preprocessor macros defined internally by the compiler. These tests
do not currently support armv7ve and so end up -mcpu= (ie an empty
string) on the compiler command line (Note: issue seen with clang,
not confirmed with gcc).

  https://gcc.gnu.org/onlinedocs/gcc-10.2.0/gcc/ARM-Options.html#ARM-Options

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>