ccsp-tr069-cacert-generic: add initial recipe
The ccsp-tr069-cacert-generic recipe can be used as a placeholder to
satisfy a runtime dependency on ccsp-tr069-cacert. It provides the
generic GlobalSign R1 Root Certificate.

A private ACS is more likely to identify itself with a certificate
signed with a proprietary Root Certificate so in production builds
the run time dependency ccsp-tr069-cacert is expected to be
satisfied by an alternative package.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>