mng-image-minimal: quote BRANCH, VERSION and SPIN values in /version.txt
Based on review of the current RDKB code, only the imagename and
BUILD_TIME fields in /version.txt are actually used, so cleaning up
the formatting of BRANCH, VERSION and SPIN should be safe.

The eventual goal is that shell scripts etc needing values from
/version.txt should be able to safely source the file (rather than
needing to parse it with a combination of sed, grep, cut, etc).

For reference:

imagename is not quoted and uses ':' as a separator (which prevents
/version.txt from being sourced directly by a shell scripts etc).
Cleanup of the imagename field in /version.txt and the code which
parses it will be done later.

BUILD_TIME is already quoted.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>