dropbear: reenable support for ssh command on the target
Normally the ssh command (ie used to initiate outgoing connections
from the target) is not needed and can be removed to save space.

However it may be used in multi-CPU systems for communication between
two CPUs, so leave enabled by default.

Targets where it's not required may (should) disable it via (for
example) a .bbappend in a target specific meta layer.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>