busybox-inittab: add local copy of "GPL-2.0-only" license text
In OE 3.3 the meta/files/common-licenses/GPL-2.0 files was renamed to
GPL-2.0-only:

  https://git.openembedded.org/openembedded-core/commit/?id=5ecf139a31fa7bd813855f1235ea9f434fbcb2e0

References to the old "GPL-2.0" license file will fail when building
with OE 3.3 (and references to the new "GPL-2.0-only" license file
will fail with OE 3.2 and earlier).

Since this meta layer intends to be compatible with multiple versions
of OE, add a local copy of the license file to break the dependency
on finding a particular file name in oe-core.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>