ccsp-dmcli: add dependency on utopia
This has been required since Alias Manager support was added to
ccsp-dmcli (the new code calls syscfg_get() to determine whether
Alias Manager functionality should be enabled).

The need for the dependency was originally missed as it was being
satisfied indirectly via the dependency on telemetry, however it
should be explicit.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>