bash: don't install the bashbug script
The bashbug script is a tool to send emails to the Bash developers
with information (build options, etc) which may be useful to help
investigate bugs in bash. It's not appropriate for embedded devices
(especially ones which use the older GPLv2 release of bash rather
than the latest GPLv3 version).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>