local.conf.sample: set more conservative BB_DISKMON_DIRS limits
By default the difference between the STOPTASKS and ABORT limits is
small enough that if the STOPTASKS limits are hit it's likely that
the ABORT limits will be hit soon after. Setting lower limits for
STOPTASKS (lower limit for disk usage == higher value for free space
required) should increase the chance that the build will be able to
stop gracefully and avoid triggering the ABORT limits.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>