libupnp: [CallStranger] Restrict UPnP SUBSCRIBE/NOTIFY traffic to the GW private IP network
The UPnP server MUST NOT accept a SUBSCRIBE message with CALLBACK value
that is not on the LAN network. The solution is:
1. When received SUBSCRIBE message,  check if source IP of SUBSCRIBE
message is private IP address
private ip address range is:
  10.0.0.0 - 10.255.255.255 (10/8 prefix)
  172.16.0.0 - 172.31.255.255 (172.16/12 prefix)
  192.168.0.0 - 192.168.255.255 (192.168/16 prefix)
    - if yes, go ahead
    - if No, return HTTP_PRECONDITION_FAILED
2. UPnP will parse and resolve callback URL
3. Check the parsed IP lists, if callback IP == source ip
    - if yes, go ahead
    - if No, return HTTP_PRECONDITION_FAILED

TESTING PERFORMED:
    Validate:
    - sending correct CALLBACK IP is accepted
    - sending incorrect IP is rejected with "Precondition Failed"
    - sending several CALLBACK IP is accepted if all are requester IP
    - sending several CALLBACK IP is rejected with "Precondition Failed"
      if one IP does not meet the criteria

Note:
If the callback IP is different than the source IP,
then the notify events are sent to the callback IP
which can lead to an amplification attack.
private-nets_1.6.22.patch doesn't cover this scenario

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>