dibbler: fix erouter0 not getting ipv6 address in dualstack mode ( OFW-1677 )
[RCA]
Issue is due to timing mismatch. dibbler-client triggered at the time when the IPv6 link local is still in tentative state.
In this case dibbler-client exits with an error saying "Interface erouter0/12 has link-local address but it is currently tentative."
Restarting dibbler-client would help to obtain the IP.
adding inactive-mode in client.conf as a fix
In inactive-mode client will accept inactive interfaces, will add them to inactive list and will periodically monitor its state.
When the interface goes on-line, client will configure it

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)