dnsmasq: DAD mechanism behavior before replying with DHCP ACK ( OFW-1148 )
When DAD is enabled, if the gateway receives a DHCP Request(renew) for an IP address instead of a DHCP Discover,
then before replying with a DHCP ACK, the gateway should broadcast an ARP query for the IP address
that it is about to Acknowledge.

The gateway should send only one ARP query and with a timeout of 250ms.
If an ARP reply is received from another client, other than the one requesting the IP address,
then the gateway should reject the DHCP Request and reply with a DHCP NAK.

If after 250ms no ARP reply has been received, then the gateway may proceed with the DHCP ACK.

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>