dnsmasq: DAD mechanism before replying with DHCP Offer ( OFW-1147 )
[Description]

When DAD is enabled, if the gateway receives a DHCP Discovery from a client device, then before replying with a DHCP Offer,
it must first broadcast an ARP query for the IP address that it is about to offer. The gateway must send
only 1 (one) ARP query and with a timeout of 250ms. If an ARP reply is received within 250ms,
then the gateway MUST select a new IP from the list of available addresses and resend a new ARP. (loop till NO ARPs are received)
If after 250ms NO ARP reply has been received, then the gateway may proceed with the DHCP Offer.

[Solution]

1. Before offering an IP, send an ARP request to that IP, and if an ARP reply is received then add it to a linked list.
2. Pass this linked list to address_allocate() api and select an IP which is not part of this linked list.
3. Repeat step 1 and 2 with the new IP till no ARP replies are received.

[Testing Performed]

1. Connect three lan clients
2. Assign two consecutive IPs to lan1 and lan2 (usually, dnsmasq offer IPs consecutively)
3. Run DHCP client on lan3
4. Capture the packets on gw and verify the ARP request before DHCP offer
5. lan3 should get a unique IP from DHCP offer (should not be the same as lan1 or lan2)

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>