dibbler: return NotOnLink if client sends request with invalid link address ( OFW-2360 )
Fix status code Success(0) was sent instead of NotOnLink(4) while
initiating a DHCPv6 Request with unknown IAID.

CDRouter test case dhcpv6_server_22

[Issue]
When Initiate a DHCPv6 Request from the client with an unknown IAID
containing an unknown address, which is not considered "on link", the
server should send a Reply message with a status code of 4 'NotOnLink'.
The current behavior is that unexpected status code Success (0) is
sent instead of  NotOnLink (4)

[Solution]
Add specific handling of the usecase described above to dibbler.

Signed-off-by: obenabdallah@libertyglobal.com