dibbler: create /var/log/dibbler from dibbler-init.sh ( OFW-2691 )
The /var/log/dibbler directory may not be created in all cases before
dibbler is started (e.g. it's currently missing if rdk-wanmanager is
used for provisioning). If it's not found then dibbler (both server
and client) does not create log files. Create the directory from
dibbler-init.sh to ensure it's always there (and to align with the
upstream RDKB version of dibbler-init.sh).

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>