quagga: don't add jitter to interval for sending RIP responses ( OFW-2682 F3896SI-1012 )
Fix CPE advertisement of RIP routes does not always match with the Update Interval time.

BUG DESCRIPTION:

When CPE is sending the RIP response, the time is deviating with the update interval for every message.
This is because of some amount of jitter is added to the update time while sending the RIP response.

FIX DESCRIPTION:

Changes are made to align with the MVXREQ-1781, to update the RIP response with Update Interval time.

TESTING PERFORMED:

1. Enabled the RIP feature.
2. Ensure  GPV of Device.Routing.RIP.InterfaceSetting.{i}.X_LGI-COM_UpdateInterval should return '30'sec - default value.
3. Started the packet capture to capture packets between eRouter and CMTS.
4. Verified that CPE is sending RIPv2 route advertisements every 30 seconds.
5. The above behaviour is tested with different update intervals.

Signed-off-by: Narayana DVVS <ndvvs.contractor@libertyglobal.com>