busybox: set initial system time from within /etc/rcS
Normally in an OE system, transfering /etc/timestamp to the system
time is done from bootmisc.sh from the initscripts package. However
we don't currently include that in OFW images, so it' not done by
default.

For Puma 6 ARM platforms the initial system time was being set from
the sdkinit script. Add to rcS so that all platforms are supported.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>