libxml2: really remove dependency on python3
The attempt to conditionally inherit a python related class in the
libxml2 recipe based on PACKAGECONFIG doesn't work as expected (the
class is always inherited even when python is being removed from
PACKAGECONFIG). See:

  https://lore.kernel.org/all/aebe57d858d895a9266a9a25e90b60c8a34dd9e3.camel@redrectangle.org/T/

As a workaround, try to undo the effects of the inherited class by
removing python3 from EXTRA_PYTHON_DEPENDS.

Note that this issue is only seen with OE 3.1 and above. In OE 3.0
the python class being inherited is related to the host (which was
apparently a bug, fixed in OE 3.1) so doesn't add a dependency for
the target build...

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>