dnsmasq: add locking to leases file updates ( OFW-3981 )
Fix POD remains connected via 2,4GHz backhaul after SSID/password change

[Issue]
When the mesh-agent copies the /nvram/dnsmasq.leases file from the ARM
side to the Atom side, on occasions the received file is empty or
incomplete. This can cause opensync to remove the pod's GRE tunnel if
the pod's backhaul IP (169.254.x.y) is not in the lease file which
breaks connectivity.

[Solution]
Lock the lease file in dnsmasq whenever it's updated, and use the same
lock when reading the file from the mesh-agent.

[Testing performed]
Change the dnsmasq configuration file to use a shorter lease time for the LAN and backhaul SSIDs.
Execute a dhcp release/renew continuously on a LAN client
After a long running test, verify that the GRE tunnels (pgdx-y) have not been removed and readded.