dnsmasq: fix unable to login into GUI after power loss with 40+ client devices connected ( OFW-4255 )
[Issue]
dnsmasq sends an ARP request and waits 250ms for a response (as part
of DAD) for each DISCOVER/REQUEST sent by clients. When a big number
of devices tries to acquire an IP at the same time, dnsmasq starts
sending DHCP replies after a large delay (~30 seconds or more)
because of the large backlog of incoming requests. In this case,
clients stop accepting replies because the transaction ID is no
longer valid on the client side.

[Fix]
1) Reduce the buffer size for the dhcp socket, so that requests get
   dropped after crossing a certain threshold of concurrent DHCP
   requests. This significantly reduces the maximum latency to handle
   requests. Clients whose requests have been dropped will retry
   later and receive an IP.

2) Handle an issue with DAD that causes dnsmasq to send a NAK to a
   client if it sends a DISCOVER for an IP with the IP already
   assigned to its interface.

[Testing performed]
Connect 60 clients at the same time
Verify that they all receive IP addresses.
Try to connect other clients and verify that they also receive IP addresses.