dibbler: fix LAN clients loses IPv6 connectivity when the device is in idle for more than 1 day ( OFW-4880 )
[Issue]
When dibbler-client attempts to renew the IAPD prefix and the IANA address in two separate exchanges with the WAN DHCPv6 server instead of both the IANA address and IAPD prefix at once, dibbler-server is killed.

[Solution]
Only set the ipv6_prefix sysevent when a prefix is received. Setting it to "::/0" will cause dibbler-server to be killed.
This was already the case before the changes made for OFW-2175.

[Testing]
Configure gateway in dual-stack mode
Verify that dibbler-server is running
Remove the pd part from /etc/dibbler/client.conf
restart dibbler-client
Verify afterwards that dibbler-server is still running